# apps/server/src/app.ts
import Fastify, { type FastifyRequest, type FastifyReply } from 'fastify';
import cors from '@fastify/cors';
import fastifySwagger from '@fastify/swagger';
import fastifySwaggerUi from '@fastify/swagger-ui';
import rateLimit from '@fastify/rate-limit';
import helmet from '@fastify/helmet';
import fastifyJwt from '@fastify/jwt';
import fastifyCookie from '@fastify/cookie';
import { config } from './config.js';
import { authRoutes } from './routes/auth.js';
import { userRoutes } from './routes/users.js';
import { roleRoutes } from './routes/roles.js';
import { tenantRoutes } from './routes/tenants.js';
import { permissionRoutes } from './routes/permissions.js';
import { auditRoutes } from './routes/audit.js';
import { statsRoutes } from './routes/stats.js';
import { healthRoutes } from './routes/health.js';
import { metricsRoutes } from './routes/metrics.js';
import { setupRoutes } from './routes/setup.js';
import { setupGuard } from './middleware/setup-guard.js';
import { oauthRoutes } from './routes/oauth.js';
import { samlRoutes } from './routes/saml.js';
import { scimRoutes } from './routes/scim.js';
import { webauthnRoutes } from './routes/webauthn.js';
import { optionsRoutes } from './routes/options.js';
import { clientRoutes } from './routes/clients.js';
import { resolveCorsOrigin } from './cors.js';
import { getRedis } from './utils/redis.js';
import { buildOidcProvider } from './oidc/provider.js';
import { OidcClientManager, ApiKeyManager, hashApiKey } from '@accessbase/identity';
import { apiKeysRoutes, getApiKeyManager } from './routes/api-keys.js';
import { DEFAULT_TENANT } from './utils/constants.js';
import { registerInteractionRoutes } from './oidc/interaction.js';
import { createOidcRateGuard } from './oidc/rate-guard.js';
import { existsSync } from 'node:fs';
import { resolve } from 'node:path';
import type { IncomingMessage } from 'node:http';
import fastifyStatic from '@fastify/static';
import { createAuditMiddleware, defaultAuditConfig } from '@accessbase/audit';
import type { AuditStorage } from '@accessbase/audit';

interface BuildAppOptions {
  /** Audit storage override; default = drizzle-backed PG audit_logs table. */
  auditStorage?: AuditStorage;
}

export async function buildApp(options: BuildAppOptions = {}) {
  const app = Fastify({
    // Batch P W1-5: TRUST_PROXY must actually reach Fastify — without this
    // option every IP-keyed defense (rate limit, lockout blacklist, audit ip)
    // collapses onto the proxy address in containerized deployments. Opt-in
    // (default false): enabling it REQUIRES a sanitizing reverse proxy —
    // a client-supplied X-Forwarded-For is then trusted as the socket peer.
    trustProxy: config.trustProxy,
    logger: {
      level: config.logLevel,
      transport:
        config.nodeEnv === 'development'
          ? {
              target: 'pino-pretty',
              options: { translateTime: 'HH:MM:ss Z', ignore: 'pid,hostname' },
            }
          : undefined,
    },
    requestIdHeader: 'x-request-id',
    genReqId: () => crypto.randomUUID(),
  });

  // --- Plugins ---

  await app.register(cors, {
    origin: resolveCorsOrigin(),
    credentials: true,
  });

  // W3-4 (F15a): unauthenticated /docs ships the entire API map — dev/test only.
  if (config.nodeEnv !== 'production') await app.register(fastifySwagger, {
    openapi: {
      info: {
        title: 'AccessBase API',
        description: 'AccessBase 基石层 API 文档',
        version: '1.0.0',
      },
      servers: [{ url: `http://localhost:${config.port}`, description: '开发环境' }],
      components: {
        securitySchemes: {
          bearerAuth: {
            type: 'http',
            scheme: 'bearer',
            bearerFormat: 'JWT',
          },
        },
      },
    },
  });

  if (config.nodeEnv !== 'production') await app.register(fastifySwaggerUi, {
    routePrefix: '/docs',
  });

  // --- Security middleware ---
  await app.register(helmet);

  const redis = await getRedis();
  await app.register(rateLimit, {
    max: 100,
    timeWindow: '1 minute',
    // @fastify/rate-limit v9 builds a RedisStore internally; null → plugin falls back to in-memory
    redis: redis ?? undefined,
    // Fail open: without this, a redis outage rethrows on every request → global 500s.
    // skipOnError falls back to local counting, matching SessionManager's fail-soft redis discipline.
    skipOnError: true,
  });
  await app.register(fastifyCookie);

  await app.register(
    fastifyJwt,
    config.jwtPrivateKeyPath && config.jwtPublicKeyPath
      ? await (async () => {
          const { readFileSync } = await import('node:fs');
          const privateKey = readFileSync(resolve(config.jwtPrivateKeyPath), 'utf-8');
          const publicKey = readFileSync(resolve(config.jwtPublicKeyPath), 'utf-8');
          return {
            secret: { public: publicKey, private: privateKey },
            sign: { algorithm: 'RS256' as const, expiresIn: '15m' },
          };
        })()
      : {
          secret: config.jwtSecret,
          sign: { expiresIn: '15m' },
        },
  );

  // --- Auth decorator ---

  app.decorate('authenticate', async (request: FastifyRequest, reply: FastifyReply) => {
    // API-key dual-read: ab_-prefixed bearer tokens are API keys, not JWTs —
    // route them to a sha256 lookup before jwtVerify rejects the format.
    const authHeader = request.headers.authorization;
    const bearer = authHeader?.startsWith('Bearer ')
      ? authHeader.slice('Bearer '.length)
      : undefined;
    if (bearer?.startsWith('ab_')) {
      const key = await getApiKeyManager().findByHash(hashApiKey(bearer));
      if (!key || key.revokedAt !== null || ApiKeyManager.isExpired(key.expiresAt)) {
        reply.status(401).send({
          success: false,
          error: { code: 'AUTH_001', message: 'Invalid API key' },
        });
        return;
      }
      (request as FastifyRequest & { user?: unknown }).user = {
        sub: key.id,
        type: 'apikey',
        scopes: key.scopes ?? ['*'],
        tenantId: key.tenantId,
      };
      // R7: apikey rows carry their own tenant — no DEFAULT fallback.
      request.tenantId = key.tenantId;
      return;
    }
    try {
      await request.jwtVerify();
    } catch {
      reply.status(401).send({
        success: false,
        error: { code: 'AUTH_001', message: 'Missing or invalid token' },
      });
      return;
    }
    // P0 disabled-user enforcement: re-check the status claim on every
    // authenticated request. Legacy tokens without the claim pass — 15m TTL
    // retires them naturally.
    const claims = request.user as { status?: string };
    if (claims.status && claims.status !== 'active') {
      reply.status(403).send({
        success: false,
        error: { code: 'AUTH_004', message: 'Account suspended' },
      });
      return;
    }
    // G/R3: resolve tenant from the request context. JWT claim wins;
    // legacy claim-less tokens fall back to the default tenant.
    const jwtPayload = request.user as { tenantId?: string };
    request.tenantId = jwtPayload.tenantId ?? DEFAULT_TENANT;
  });

  // --- Error envelope enrichment (security.md 19.13 / D52) ---
  app.setErrorHandler((error, request, reply) => {
    const statusCode = error.statusCode ?? 500;
    const payload: Record<string, unknown> = {
      success: false,
      error: {
        code:
          error.code === 'FST_ERR_VALIDATION'
            ? 'VALIDATION_001'
            : (error.code ?? `HTTP_${statusCode}`),
        message: statusCode >= 500 ? 'Internal server error' : error.message,
      },
      timestamp: new Date().toISOString(),
      requestId: request.id,
      path: request.url,
    };
    if (statusCode < 500) request.log.info({ err: error }, 'Request error');
    else request.log.error({ err: error }, 'Internal error');
    return reply.status(statusCode).send(payload);
  });

  // --- OIDC provider (mounted at /oidc, unauthenticated protocol endpoints) ---
  // B1/B2: Fastify parses bodies BEFORE handlers and provider.callback() is a
  // Koa factory — hijack on onRequest (before body parsing) and hand the raw
  // req/res to the cached provider handler. No content-type parsers for /oidc.
  // Registered BEFORE setupGuard: hijacked /oidc requests skip the guard's
  // DB round-trip entirely (the provider's adapter dials PG only when a
  // flow actually needs a Client/Grant lookup).
  const { createDb } = await import('@accessbase/identity/db');
  const { provider: oidcProvider, oidcHandler, stopSweeper } = await buildOidcProvider({
    issuer: `${config.oauthRedirectBase}/oidc`,
    jwtSecret: config.jwtSecret,
    nodeEnv: config.nodeEnv,
    privateKeyPath: config.jwtPrivateKeyPath,
    publicKeyPath: config.jwtPublicKeyPath,
    adapterCtorArgs: [createDb(config.databaseUrl)],
    frontendOrigin: config.frontendOrigin,
  });
  // Batch N: stop the adapter sweeper on graceful close.
  app.addHook('onClose', async () => stopSweeper());

  // Q2a(E): retention sweeper on its OWN pool (rev.2 R7: first pass boot+60s,
  // then every 24h — deploy cycles faster than a day must still sweep). The
  // audit archive config finally gets a consumer; AUDIT_RETENTION_DAYS overrides, 0 disables.
  const { startRetentionSweeper, resolveRetentionDays } = await import('./utils/retention-sweeper.js');
  const { resetManagers } = await import('./utils/managers.js');
  // Lazy pool: created on the first tick (boot+60s), closed by stop() — keeps
  // buildApp createDb-count at zero (health-pool.test premise).
  const retention = startRetentionSweeper(
    () => createDb(config.databaseUrl),
    resolveRetentionDays(
      process.env['AUDIT_RETENTION_DAYS'],
      defaultAuditConfig.storage.archive.retentionDays,
    ),
    app.log,
  );
  // Q2c: cross-node permission-cache coherence (silent single-node mode if Redis
  // is absent or the identity lane is mocked — never a boot failure).
  const { setupCacheCoherence } = await import('./utils/cache-coherence.js');
  const coherence = await setupCacheCoherence(app.log);
  // Q2a(B): graceful close ends the singleton managers' pools + the sweeper's pool.
  app.addHook('onClose', async () => {
    await coherence.teardown();
    await retention.stop();
    await resetManagers();
  });
  // Batch P W2-1: the hijacked /oidc/* space lives in Fastify's route-less
  // (404) region where @fastify/rate-limit provably never engages — coarse
  // per-IP guard for it. Exempt: interaction routes (real, globally limited)
  // and discovery/jwks (anonymous RP polls).
  const oidcGuard = createOidcRateGuard(config.oidcIpRatePerMin, redis);
  app.addHook('onRequest', (req, reply, done) => {
    // Interaction contract endpoints (Task 4c) are real Fastify routes — they
    // need body parsing + bearer auth, so they bypass the provider hijack.
    if (req.url.startsWith('/oidc/interaction/')) return done();
    if (!req.url.startsWith('/oidc/')) return done();
    // Discovery + JWKS are anonymous RP polls: hand off UNCOUNTED (they stay
    // with the provider — this is a guard exemption, not a hijack exemption).
    const handoff = (): void => {
      reply.hijack();
      // Provider routes are registered WITHOUT the /oidc prefix (issuer path =
      // mountPath); per panva's official mount docs, strip the prefix and keep
      // originalUrl so urlFor recomposes absolute URLs with the prefix.
      const raw = req.raw as IncomingMessage & { originalUrl?: string };
      raw.originalUrl = raw.url;
      raw.url = (raw.url ?? '').slice('/oidc'.length);
      oidcHandler(raw, reply.raw).then(() => done(), done);
    };
    if (req.url.startsWith('/oidc/.well-known')) return handoff();
    const denied = (): void => {
      void reply
        .code(429)
        .header('retry-after', '60')
        .type('application/json')
        .send({ error: 'invalid_request', error_description: 'rate limited' });
    };
    // The guard self-degrades to local counting on redis errors; a rejection
    // here is a programmer error — hand off uncounted rather than 500 the
    // protocol surface (skipOnError parity with the global limiter).
    void oidcGuard(req.ip).then((ok) => (ok ? handoff() : denied()), handoff);
  });

  // --- Setup Guard Middleware (must be registered before other routes) ---
  app.addHook('onRequest', setupGuard);

  // --- Audit middleware (Phase 6a Task 5) ---
  // onResponse hook on /api/v1/* write requests; storage: injected override (tests) or drizzle-backed.
  // Non-test builds without override lazily create a drizzle db — test env without override never touches PG.
  if (options.auditStorage || config.nodeEnv !== 'test') {
    const { AuditLogger } = await import('@accessbase/audit');
    let storage = options.auditStorage;
    if (!storage) {
      const { createDb, auditLogs } = await import('@accessbase/identity/db');
      const db = createDb(config.databaseUrl);
      storage = {
        async write(entries) {
          if (entries.length === 0) return;
          await db.insert(auditLogs).values(
            entries.map((e) => ({
              tenantId: e.tenantId,
              userId: e.userId,
              action: e.action,
              resourceType: e.resourceType,
              resourceId: e.resourceId,
              requestBody: e.requestBody,
              responseStatus: e.responseStatus,
              requestId: e.requestId,
              ip: e.userIp,
              userAgent: e.userAgent,
            })),
          );
        },
      };
    }
    const auditLogger = new AuditLogger(
      {
        ...defaultAuditConfig,
        level: 'write',
        async: { ...defaultAuditConfig.async, enabled: false },
      },
      { storage },
    );
    const auditHook = createAuditMiddleware(auditLogger);
    app.addHook('onResponse', async (request, reply) => {
      const url = request.url.split('?')[0] ?? '';
      if (request.method === 'GET' || request.method === 'HEAD') return;
      if (
        url.startsWith('/health') ||
        url.startsWith('/metrics') ||
        url.startsWith('/api/v1/setup') ||
        url.startsWith('/api/v1/options') ||
        // R10 (Batch F): the ACS urlencoded body (multi-KB SAMLResponse XML)
        // must not land in audit_logs — bloat + assertion content at rest.
        url.startsWith('/api/v1/auth/saml/acs')
      )
        return;
      await auditHook(request, reply);
    });
  }

  // --- Routes ---
  await app.register(healthRoutes, { prefix: '/health' });
  // L-M D2: root-level /metrics. Registered AFTER the OIDC hijack hook so
  // hijacked /oidc replies (which skip onResponse) never touch in-flight
  // accounting — documented blind spot, see routes/metrics.ts.
  await app.register(metricsRoutes);
  await app.register(setupRoutes, { prefix: '/api/v1/setup' });
  await app.register(authRoutes, { prefix: '/api/v1/auth' });
  await app.register(userRoutes, { prefix: '/api/v1/users' });
  await app.register(roleRoutes, { prefix: '/api/v1/roles' });
  await app.register(tenantRoutes, { prefix: '/api/v1/tenants' });
  // SCIM 2.0 (Batch H) — own bearer preHandler inside the plugin; never app.authenticate.
  await app.register(scimRoutes, { prefix: '/api/v1/scim/v2' });
  await app.register(permissionRoutes, { prefix: '/api/v1/permissions' });
  await app.register(auditRoutes, { prefix: '/api/v1/audit-logs' });
  await app.register(statsRoutes, { prefix: '/api/v1' });
  await app.register(optionsRoutes, { prefix: '/api/v1' });
  await app.register(clientRoutes, { prefix: '/api/v1' });
  await app.register(oauthRoutes, { prefix: '/api/v1/auth' });
  await app.register(samlRoutes, { prefix: '/api/v1/auth' });
  await app.register(webauthnRoutes, { prefix: '/api/v1/auth' });
  await app.register(apiKeysRoutes, { prefix: '/api/v1/auth/api-keys' });
  // Client display names live in the registry (the provider's Client wrapper
  // drops non-schema fields), so the interaction GET resolves them here.
  const oidcClientManager = new OidcClientManager(createDb(config.databaseUrl));
  await app.register(registerInteractionRoutes, {
    prefix: '/oidc',
    provider: oidcProvider,
    clientNameLookup: async (clientId) => (await oidcClientManager.get(clientId))?.name,
  });
  // --- L0 package registration (when packages are implemented) ---
  // await app.register(identityPlugin)
  // await app.register(auditPlugin)
  // await app.register(healthCheckPlugin)
  // await app.register(i18nPlugin)

  // --- Static file serving (deploy mode) ---
  if (existsSync(resolve(config.staticDir))) {
    await app.register(fastifyStatic, {
      root: resolve(config.staticDir),
      prefix: '/',
      index: ['index.html'],
      decorateReply: true,
      wildcard: false,
    });
  }
  // --- SPA fallback / 404 envelope (works with or without static files) ---
  app.setNotFoundHandler(async (request, reply) => {
    if (
      request.url.startsWith('/api/') ||
      request.url.startsWith('/health') ||
      request.url.startsWith('/docs')
    ) {
      return reply.status(404).send({
        success: false,
        error: { code: 'NOT_FOUND', message: 'Resource not found' },
        timestamp: new Date().toISOString(),
        requestId: request.id,
        path: request.url,
      });
    }
    // SPA fallback: serve index.html for non-API routes
    if (existsSync(resolve(config.staticDir))) {
      return reply.type('text/html').sendFile('index.html');
    }
    return reply.status(404).send({
      success: false,
      error: { code: 'NOT_FOUND', message: 'Resource not found' },
      timestamp: new Date().toISOString(),
      requestId: request.id,
      path: request.url,
    });
  });

  return app;
}